+++ Source/ACE/ACE.csproj
@@ -122,7 +122,6 @@
     <Compile Include="Entity\Model.cs" />
     <Compile Include="Entity\ModelPalette.cs" />
     <Compile Include="Entity\ModelTexture.cs" />
-    <Compile Include="Entity\ModelData.cs" />
     <Compile Include="Entity\Portal.cs" />
     <Compile Include="Entity\SpellLikeEffect.cs" />
     <Compile Include="Entity\SummonedPortal.cs" />
