--- Source/ACE/Entity/ModelData.cs
@@ -1,98 +1,0 @@
-﻿using ACE.Network;
-using ACE.Network.Enum;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-
-namespace ACE.Entity
-{
-    /// <summary>
-    /// Segment to Control AC Model / Palettes and Textures.
-    /// TODO: remove and read directly from WorldObject.AceObject
-    /// </summary>
-    public class ModelData
-    {
-        public uint? PaletteGuid { get; set; } = 0;
-
-        public uint? ClothingBase { get; set; } = 0;
-
-        private readonly List<ModelPalette> modelPalettes = new List<ModelPalette>();
-
-        private readonly List<ModelTexture> modelTextures = new List<ModelTexture>();
-
-        private readonly List<Model> models = new List<Model>();
-
-        public List<ModelPalette> GetPalettes
-        {
-            get { return modelPalettes.ToList(); }
-        }
-
-        public List<ModelTexture> GetTextures
-        {
-            get { return modelTextures.ToList(); }
-        }
-
-        public List<Model> GetModels
-        {
-            get { return models.ToList(); }
-        }
-
-        public void AddPalette(uint paletteID, ushort offset, ushort length)
-        {
-            ModelPalette newpalette = new ModelPalette(paletteID, offset, length);
-            modelPalettes.Add(newpalette);
-        }
-
-        public void AddTexture(byte index, uint oldtexture, uint newtexture)
-        {
-            ModelTexture nextTexture = new ModelTexture(index, oldtexture, newtexture);
-            modelTextures.Add(nextTexture);
-        }
-
-        public void AddModel(byte index, uint modelresourceid)
-        {
-            var newmodel = new Model(index, modelresourceid);
-            models.Add(newmodel);
-        }
-
-        public void Clear()
-        {
-            modelPalettes.Clear();
-            modelTextures.Clear();
-            models.Clear();
-        }
-
-        // todo: render object network code
-        public void Serialize(BinaryWriter writer)
-        {
-            writer.Write((byte)0x11);
-            writer.Write((byte)modelPalettes.Count);
-            writer.Write((byte)modelTextures.Count);
-            writer.Write((byte)models.Count);
-
-            if ((modelPalettes.Count > 0) && (PaletteGuid != null))
-                writer.WritePackedDwordOfKnownType((uint)PaletteGuid, 0x4000000);
-            foreach (var palette in modelPalettes)
-            {
-                writer.WritePackedDwordOfKnownType(palette.PaletteId, 0x4000000);
-                writer.Write((byte)palette.Offset);
-                writer.Write((byte)palette.Length);
-            }
-
-            foreach (var texture in modelTextures)
-            {
-                writer.Write((byte)texture.Index);
-                writer.WritePackedDwordOfKnownType(texture.OldTexture, 0x5000000);
-                writer.WritePackedDwordOfKnownType(texture.NewTexture, 0x5000000);
-            }
-
-            foreach (var model in models)
-            {
-                writer.Write((byte)model.Index);
-                writer.WritePackedDwordOfKnownType(model.ModelID, 0x1000000);
-            }
-
-            writer.Align();
-        }
-    }
-}
